project/requirements.txt
Django==3.0.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
six==1.14.0
Pillow==7.1.2
boto3==1.13.20
django-storages==1.9.1
django-cors-headers==3.3.0
psycopg2-binary==2.8.5
psycopg2==2.8.5
s3transfer==0.3.3
django-crontab==0.7.1